Add support for making shortcuts to emulated games
Currently uses a horrible reflection hack to access the
EmulatorProfileExtensions.ExpandVariables() method. This method is
actually an extension method to EmulatorProfile, but we must invoke them
directly as static methods in IronPython. Furthermore, the
EmulatorProfileExtensions class isn't exposed to the extension scripting
environment, so we use reflection to access it anyway.